:ok_hand: Implemented required changes (Udacity's project rubric guidelines)
The sectionName is now shown. It is required as per Udacity's project rubric guidelines for this exercise. In my opinion, in the scope of this app it is not necessary as the sectionName is the same for every single articles falling in the category I chose: they're pieces of "Opinion" (remember the infamous "every article author is Editorial?" be5e20217b654e7a0a547c57211181c402b472fb Well here we go again)

Keeping it separated from master as I don't want to merge those changes (but I do want to validate my submission 😁)